+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# Version x.x.x
+- Add usage to resetdb command in wordpress module.
+- Add PHP-Nginx-MySQL module
 # Version 1.0.0
 - Initial Release